+++ build.sh
@@ -134,6 +134,18 @@ fi
 ./wireguard/build-wireguard-go.sh
 
 echo "Building Rust code in release mode using $RUSTC_VERSION..."
+
+if [[ ("$(uname -s)" == "Darwin") || ("$(uname -s)" == "Linux") ]]; then
+    pushd mullvad-cli
+    mkdir -p "$SCRIPT_DIR/dist-assets/shell-completions"
+    for sh in bash zsh; do
+        echo "Generating shell completion script for $sh..."
+        cargo +stable run $CARGO_ARGS --release --features shell-completions -- \
+            shell-completions "$sh" "$SCRIPT_DIR/dist-assets/shell-completions/"
+    done
+    popd
+fi
+
 MULLVAD_ADD_MANIFEST="1" cargo +stable build $CARGO_ARGS --release
 
 ################################################################################
@@ -197,7 +209,6 @@ fi
 
 
 ./update-relays.sh
-
 
 pushd "$SCRIPT_DIR/gui"